Move device-tree parsing into a separate crate
