+++ crates/device-tree/Cargo.toml
@@ -1,12 +1,7 @@
 [package]
-name = "kernel"
+name = "device-tree"
 version = "0.1.0"
 edition = "2021"
 
-[[bin]]
-name = "kernel"
-path = "src/main.rs"
-
 [dependencies]
 bytemuck = { version = "1.19.0", default-features = false }
-device-tree = { path = "../device-tree" }
